pyproject.toml
[project]
name = "ai-news-agg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "feedparser>=6.0.12",
    "google-genai>=1.0.0",
    "pydantic-settings>=2.0.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "sqlalchemy>=2.0.48",
    "youtube-transcript-api>=1.2.4",
]
